refactor(app): move filterProfanity from shared-util to Application class
this function relies on a main object that is passed through a parameter. It is like creating a static function and then passing "this" as a parameter.

To fix the issue, the function is moved into the object that directly hosts the profanityFilter object.
