+++ src/util/shared-util.ts
@@ -1,4 +1,3 @@
-import type { Filter } from 'bad-words'
 import log4js from 'log4js'
 
 export function sufficeToTime(suffice: string): number {
@@ -63,24 +62,4 @@ export const escapeDiscord = function (message: string): string {
   message = message.split('@').join(String.raw`\@-`) // mentions
 
   return message
-}
-
-export function filterProfanity(
-  playerMessage: string,
-  profanityFilter: Filter
-): { filteredMessage: string; changed: boolean } {
-  let filtered: string
-  try {
-    filtered = profanityFilter.clean(playerMessage)
-  } catch {
-    /*
-        profanity package has bug.
-        will throw error if given one special character.
-        example: clean("?")
-        message is clear if thrown
-      */
-    filtered = playerMessage
-  }
-
-  return { filteredMessage: filtered, changed: playerMessage !== filtered }
 }
